Removed unused scripts and other plotter improvements
Removed unused scripts from rogueinabox.
Fix: When the training starts or it is stopped and then restarted again, the first x statistics are too noisy because they are calculated using less than x episodes. What this fix does is to ignore the first x episodes, thus not printing them in the log file.